ce,ee allow create public dataset
