+++ api/handler/dataset_ce.go
@@ -1,0 +1,7 @@
+//go:build !saas
+
+package handler
+
+func (h *DatasetHandler) allowCreatePublic() bool {
+	return true
+}
